+++ engine/api/golang/kurtosis_engine_rpc_api_consts/kurtosis_engine_rpc_api_consts.go
@@ -1,9 +1,0 @@
-package kurtosis_engine_rpc_api_consts
-
-const (
-	// Protocol that the engine server container listens on
-	ListenProtocol = "tcp"
-
-	// The port that the engine server container listens on
-	ListenPort uint16 = 9710
-)