+++ engine/api/typescript/src/index.ts
@@ -1,7 +1,6 @@
 export { KURTOSIS_ENGINE_API_VERSION } from "./kurtosis_engine_api_version/kurtosis_engine_api_version";
-export { LISTEN_PORT, LISTEN_PROTOCOL } from "./kurtosis_engine_rpc_api_consts/kurtosis_engine_rpc_api_consts";
 
-export { KurtosisContext } from "./lib/kurtosis_context/kurtosis_context";
+export { KurtosisContext, DEFAULT_KURTOSIS_ENGINE_SERVER_PORT_NUM } from "./lib/kurtosis_context/kurtosis_context";
 
 // RPC API bindings
 export { EngineServiceClient } from "./kurtosis_engine_rpc_api_bindings/engine_service_grpc_pb";