+++ engine/api/typescript/src/kurtosis_engine_rpc_api_consts/kurtosis_engine_rpc_api_consts.ts
@@ -1,5 +1,0 @@
-// Protocol that the engine server container listens on
-export const LISTEN_PROTOCOL: string = "tcp";
-
-// The port that the engine server container listens on
-export const LISTEN_PORT: number = 9710;